+++ src/h3sed/conf.py
@@ -26,7 +26,7 @@ import sys
 """Program title, version number and version date."""
 Name = "h3sed"
 Title = "Heroes3 Savegame Editor"
-Version = "3.6.dev3"
+Version = "3.6.dev4"
 VersionDate = "20.03.2026"
 
 Frozen = getattr(sys, "frozen", False)